stats: turn loan histogram endpoint into reusable component
