+++ invenio_app_ils/stats/histogram/serializers/schema.py
@@ -6,7 +6,7 @@
 # Invenio is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 
-"""Invenio App ILS loan stats serializers schema."""
+"""Invenio App ILS histogram stats serializers schema."""
 
 from marshmallow import Schema, fields
